requirements.txt
django-environ==0.12.0
django-jazzmin==3.0.1
pillow==10.2.0
djangorestframework==3.15.2
Django==5.1.2
django-cors-headers==4.6.0